requirements.txt

openai
click
python-dotenv
black
isort
pytest
